+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - **fix:** extract #extendRouteList
 - **doc:** fix typo
 - **upg:** babel-cli, babel-core and babel-preset-2015 (6.8.0 -> 6.9.0) and eslint (2.10.1 -> 2.10.2)
+- **add:** documentation of parameter
 
 ### [v0.3.1] - 2016-05-17
 - **doc:** add explanation for options.log